Update all double values to BigDecimal
Converted all double values managing money to BigDecimal. This is for more
accurate rounding and displaying of money to the user.

+++ src/com/view/CashPaymentView.java
@@ -75,7 +75,7 @@ public class CashPaymentView extends JFrame
               lblChange.setText("Change : £" + paymentController.CalculateChange());
 
 
-              if (paymentController.cashAdded >= paymentController.cashDue)
+              if (paymentController.cashAdded.doubleValue() >= paymentController.cashDue.doubleValue())
               {
                   btnCompletePayment.setEnabled(true);
